feat(layout): migrate footer-client ts
