+++ globals.d.ts
@@ -1,1 +1,3 @@
 declare module '*.md';
+
+declare module '*.svg';
