+++ packages/Layout/footer-client/package.json
@@ -2,7 +2,7 @@
   "name": "@axa-fr/react-toolkit-layout-footer-client",
   "version": "1.4.0-alpha.0",
   "main": "dist/index.js",
-  "jsnext:main": "src/index.js",
+  "types": "dist/index.d.ts",
   "private": false,
   "scripts": {
     "prepare": "node ../../../scripts/prepare.js"
@@ -12,10 +12,7 @@
     "react-dom": "^16.8.0 || ^17.0.0"
   },
   "dependencies": {
-    "@axa-fr/react-toolkit-core": "1.4.0-alpha.0",
-    "@types/classnames": "2.2.11",
-    "classnames": "2.2.6",
-    "prop-types": "15.7.2"
+    "@axa-fr/react-toolkit-core": "1.4.0-alpha.0"
   },
   "license": "MIT",
   "publishConfig": {
